+++ next/components/moderator-control-panel.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 import ModeratorDeckEditor from './moderator-deck-editor.js';
 import ModeratorConfigEditor from './moderator-config-editor.js';
 import ModeratorFeaturedDecksEditor from './moderator-featured-decks-editor.js';
+import ModAccountEditor from './mod-account-editor.js';
 
 function moderatorControlPanel() {
   return (
@@ -12,6 +13,8 @@ function moderatorControlPanel() {
       <ModeratorConfigEditor />
       <hr />
       <ModeratorFeaturedDecksEditor />
+      <hr />
+      <ModAccountEditor />
     </>
   );
 }