+++ next/pages/account.js
@@ -1,14 +1,40 @@
-import { useContext, useState } from 'react';
+import { useContext, useEffect, useState } from 'react';
+import { withRouter } from 'next/router';
 import Layout from '../components/layout';
 import UserContext from '../components/user-context';
 import { ApolloConsumer } from 'react-apollo';
 import updateUsername from '../lib/mutations/update-username';
 import UserDecks from '../components/user-decks';
+import PublicAccount from '../components/public-account.js';
+import Profile from '../components/profile.js';
+import AvatarPicker from '../components/avatar-picker.js';
+import Router from 'next/router';
+import { useMutation } from '@apollo/react-hooks';
+import gql from 'graphql-tag';
 
-const cdn = process.env.MG_CDN;
+const error403 = () => Router.push('/');
 
-function Account() {
+export default withRouter(({ router }) => {
+  const un = router.query.name;
+  if (un) {
+    return <PublicAccount username={un} />;
+  }
+
   const { user, updateUser } = useContext(UserContext);
+
+  let result = null;
+
+  useEffect(() => {
+    if (!user) {
+      error403();
+      return;
+    }
+  });
+
+  if (!user) {
+    return <Layout>{result}</Layout>;
+  }
+
   const [username, setUsername] = useState(user.username);
 
   const areSettingsValid = () => !!username;
@@ -17,47 +43,56 @@ function Account() {
     if (!areSettingsValid()) return;
     updateUsername(apolloClient, user.id, username)
       .then(({ data }) => {
-        updateUser(data.updateAccount.account);
+        updateUser({ ...user, ...data.updateAccount.account });
       })
       .catch(err => {
+        // TODO alert special msg if username not unique (or just suggest
+        // a different one regardless of error).
         console.error('Something went wrong while updating user name', err);
       });
   };
 
-  const regDate = new Date(user.registered);
-  const regDateString = regDate.toLocaleDateString('en-us', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric'
-  });
+  const [patchAvatar] = useMutation(
+    gql`
+      mutation updateProfileIconId($accountId: Int!, $profileIconId: Int!) {
+        updateAccount(
+          input: { id: $accountId, patch: { profileIconId: $profileIconId } }
+        ) {
+          account {
+            profileIconId
+          }
+        }
+      }
+    `,
+    {
+      update() {
+        window.scrollTo(0, 0);
+        // TODO - write to cache instead?
+        window.location.reload();
+      },
+      onError() {
+        alert('That failed. Please check values and try again');
+      }
+    }
+  );
+
+  const updateAvatar = profileIconId => {
+    patchAvatar({
+      variables: {
+        accountId: user.id,
+        profileIconId
+      }
+    });
+  };
 
   return (
     <>
       <style jsx>{`
-        .user-profile {
+        .private-profile {
           text-align: center;
           width: 100%;
           border: 1px solid #458a9e;
-          background-color: #1c2d35;
         }
-
-        .profile-image {
-          width: 138px;
-          margin-top: 30px;
-        }
-
-        .user-name {
-          color: #f1810b;
-          font-size: 1.5em;
-          font-style: italic;
-          font-weight: 600;
-        }
-
-        .member-since {
-          font-style: italic;
-          font-weight: 300;
-        }
-
         .profile-content {
           display: flex;
           flex-wrap: wrap;
@@ -96,14 +131,8 @@ function Account() {
         title="Mythgard Hub | Account Settings"
         desc="Account settings for Mythgard Hub"
       >
-        <div className="user-profile">
-          <img
-            src={`${cdn}/mgh/avatar2.png`}
-            alt="Profile Icon"
-            className="profile-image"
-          />
-          <div className="user-name">{user.username}</div>
-          <div className="member-since">Member since {regDateString}</div>
+        <div className="private-profile">
+          <Profile user={user} />
           <br />
           <br />
           <div className="profile-content">
@@ -137,15 +166,21 @@ function Account() {
                     <button
                       type="submit"
                       value="Save"
-                      style={{ width: '25%' }}
+                      style={{ width: 'auto' }}
                       onClick={() => {
                         handleSubmit(client);
                       }}
                     >
-                      Save
+                      Save Username
                     </button>
                   )}
                 </ApolloConsumer>
+                <div>
+                  <AvatarPicker
+                    onSave={profileIconId => updateAvatar(profileIconId)}
+                    accountType={user.accountType}
+                  />
+                </div>
               </div>
             </div>
           </div>
@@ -153,8 +188,4 @@ function Account() {
       </Layout>
     </>
   );
-}
-
-Account.requiresAuth = true;
-
-export default Account;
+});