+++ next/server-routes/auth.js
@@ -14,14 +14,27 @@ const client = new PgClient({
   ssl: process.env.PGSSL === 'yes'
 });
 
+const camelizeUserRecord = userRecord => {
+  userRecord.profileIconId = userRecord.profile_icon_id;
+  delete userRecord.profile_icon_id;
+  userRecord.accountType = userRecord.account_type;
+  delete userRecord.account_type;
+};
+
 /**
  * Fetches an account (user) record by email.
  */
 const getUserByEmail = async email => {
-  const query = `SELECT id, email, username, registered FROM mythgard.account WHERE email = $1`;
+  const query = `SELECT id, email, username, profile_icon_id, registered, account_type FROM mythgard.account WHERE email = $1`;
   try {
     const res = await client.query(query, [email]);
-    return res.rows[0];
+    const [userRecord] = res.rows;
+    try {
+      camelizeUserRecord(userRecord);
+    } catch {
+      /* noop */
+    }
+    return userRecord;
   } catch (err) {
     return null;
   }